Move account-fingerprint to KM ownership
