+++ apps/web/src/app/key-management/account-fingerprint/account-fingerprint.component.html
